refactor: '/' 경로 로직 수정 - '/home' 추가
